# Optimize ListViewRead<BitView> => Bitlist conversion