<!-- Thanks for sending a pull request! Please check out our contribution guidelines: -->
<!-- https://github.com/ConsenSys/teku/blob/master/CONTRIBUTING.md -->

## PR Description

This rather a tweak than full-fledged views refactoring but it works well. 
Since SSZ serialization is now pretty fast (it just iterates leaf nodes), serialize-deserialize `Bitlist` works much faster than accessing individual bits via traversing all backing tree nodes every time. 

Before:
```
Benchmark                             Mode  Cnt      Score     Error  Units
BitlistBenchmark.fromCachedListView  thrpt   25  19158,888    876,506 ops/s
BitlistBenchmark.fromNewListView     thrpt   25   2514,544    123,885 ops/s
```

After:
```
BitlistBenchmark.fromCachedListView  thrpt   25  42554,508    3296,832 ops/s
BitlistBenchmark.fromNewListView     thrpt   25  40462,349    3766,117 ops/s
```

## Fixed Issue(s)

Fix #3375

## Documentation

- [x] I thought about documentation and added the `documentation` label to this PR if updates are required.