@Nashatyrev 2020-12-16
Em... it's ~x16 faster when non-cached and >x2 faster when cached.