@ajsutton 2020-12-16
> Em... it's ~x16 faster when non-cached and >x2 faster when cached.

Ah yes, I see the extra digit in there now.  In your locale is `,` used for the decimal point or is it just doing a really bad job at using comma to separate thousands?