@Nashatyrev 2020-12-16
Yep, it's a decimal separator in Ru locale and it always bothers me
(BTW just discovered interesting fact: the comma decimal separator appeared to be more widespread across countries than dot)