@Nashatyrev 2020-12-16
The failing test appears to be the Tuweni issue: https://github.com/apache/incubator-tuweni/issues/186